Your descriptive message

+++ backend/c4-backend/products/product.route.js
@@ -4,6 +4,7 @@ const { isValidObjectId } = require("mongoose");
 const userModel = require("../users/user.model");
 const validateMiddleware = require("../middlewares/validate.middleware");
 const productSchema = require("./product.dto");
+const isAuthMiddleware = require("../middlewares/is-auth.middleware");
 
 const productRouter = new Router()
 
@@ -34,15 +35,15 @@ productRouter.get('/', async (req, res) => {
     res.json(products)
 })
 
-productRouter.post('/', validateMiddleware(productSchema), async (req, res) => {
+productRouter.post('/',isAuthMiddleware, validateMiddleware(productSchema), async (req, res) => {
     const newProduct = await productModel.create({
         name: req.body.name,
         price: req.body.price,
         desc: req.body.desc,
         review: req.body.review,
-        seller: req.body.userId
+        seller: req.userId
     })
-    await userModel.findByIdAndUpdate(req.body.userId, {
+    await userModel.findByIdAndUpdate(req.userId, {
         $push: {products: newProduct._id}
     })
     res.status(201).json({success: true, message: "Created successfully"})
@@ -59,9 +60,12 @@ productRouter.get('/:id', isValidObjectId, async (req, res) => {
     res.json(product)
 })
 
-productRouter.delete('/:id', isValidObjectId, async (req, res) => {
+productRouter.delete('/:id', isAuthMiddleware , isValidObjectId, async (req, res) => {
     const id = req.params.id
-    
+    const product = await productModel.findById(id)
+    if(product.seller !== req.userId) {
+        return res.status(401).json({message: "premision denied"})
+    }
     const deletedProduct = await productModel.findByIdAndDelete(id)
     if(!deletedProduct){
         return res.status(404).json({message: "product not found"})
@@ -74,8 +78,12 @@ productRouter.delete('/:id', isValidObjectId, async (req, res) => {
     res.json(deletedProduct)
 })
 
-productRouter.put('/:id', isValidObjectId, async (req, res) => {
+productRouter.put('/:id',  isAuthMiddleware,isValidObjectId, async (req, res) => {
     const id = req.params.id
+     const product = await productModel.findById(id);
+     if (product.seller !== req.userId) {
+       return res.status(401).json({ message: "premision denied" });
+     }
 
     const updatedProduct = await productModel.findByIdAndUpdate(id, req.body, {new: true})
     if(!updatedProduct){
